+++ waspc/e2e-tests/test-outputs/snapshots/wasp-migrate-golden/wasp-app/.wasp/out/sdk/wasp/client/vite/virtual-files/files/client-entry.tsx
@@ -8,15 +8,13 @@ import { WaspApp } from "wasp/client/app";
 
 import { routeObjects } from '/@wasp/routes.tsx'
 
+// We embed this data at prerendering time.
+const { isFallbackPage, routerHydrationData } = window.__WASP_SSR_DATA__ ?? {}
+
 const router = createBrowserRouter(routeObjects, {
   basename: "/",
-  // React Router will put hydration data on this property of the `window` object.
-  // https://reactrouter.com/7.13.1/start/data/custom#4-hydrate-in-the-browser
-  hydrationData: window.__staticRouterHydrationData,
+  hydrationData: routerHydrationData,
 })
-
-// We embed this data at prerendering time.
-const { isFallbackPage } = window.__WASP_SSR_DATA__ ?? {}
 
 function App() {
   return (