+++ waspc/e2e-tests/test-outputs/snapshots/wasp-migrate-golden/wasp-app/.wasp/out/sdk/wasp/vite-env.d.ts
@@ -2,9 +2,9 @@
 
 interface WaspSSRData {
   isFallbackPage: boolean
+  routerHydrationData?: import("react-router").HydrationState
 }
 
 interface Window {
   __WASP_SSR_DATA__?: WaspSSRData
-  __staticRouterHydrationData?: import("react-router").HydrationState
 }